% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_factor_sum.R
\name{calc_factor_sum}
\alias{calc_factor_sum}
\title{Calculates factor scores by summing the variables}
\usage{
calc_factor_sum(.data, ...)
}
\arguments{
\item{.data}{A dataframe}

\item{...}{Variables to sum. Select variables DPLYR style.
One or more unquoted expressions separated by commas.
You can treat variable names like they are positions, so you can use expressions like x:y to select ranges of variables
You can also pass variables as vector of strings}
}
\value{
A vector of numeric values representing the factor scores per subject
}
\description{
This function takes a dataframe, and a series of variables and calculates the
rowwise sum. It should be usefull to calculate factor scores out of several variables
}
\details{
Beware that
* The output is a vector of numeric values. To use this in a piping syntax
see the example below - note the use of '.' in the pipe worlflow
* The sum throws out the missing values (na.rm = TRUE) so you will always get a sum despite the missing values in you variables
}
\examples{
mtcars \%>\% dplyr::mutate(score = calc_factor_sum(., mpg, cyl)) \%>\% head()
}
