% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_bar_chart.R
\name{lp_bar_chart}
\alias{lp_bar_chart}
\title{A bar chart with percentages}
\usage{
lp_bar_chart(
  data,
  var,
  .xlab = NULL,
  .ylab = NULL,
  .title = NULL,
  .bar_col = "#EE5C42"
)
}
\arguments{
\item{data}{The dataset}

\item{var}{The variable to plot}

\item{.xlab}{The x-axis label. Defaults to the variable name}

\item{.ylab}{The y-axis label. Defaults to the variable name}

\item{.title}{Optional title}

\item{.bar_col}{The colour of the bars}
}
\value{
A ggplot object
}
\description{
A bar chart with percentages
}
