% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_missing.R
\name{lp_missing}
\alias{lp_missing}
\title{Count missing observations}
\usage{
lp_missing(.data, ...)
}
\arguments{
\item{.data}{A dataframe, or an aboject that can be coerced to dataframe}

\item{...}{Unused yet}
}
\value{
A tibble
}
\description{
This function counts the number of missing observations per variable
}
\examples{
lp_missing(airquality)
}
