% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lp_short}
\alias{lp_short}
\title{Wrap a long text}
\usage{
lp_short(x, width = 30)
}
\arguments{
\item{x}{String length 1}

\item{width}{Width to wrap it. See \code{base::strwrap}}
}
\value{
String length 1
}
\description{
Wrap a long text
}
\examples{
lp_short("A very long string that needs to take more than one lines")
}
