% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_summarise.R
\name{lp_summarise}
\alias{lp_summarise}
\title{Summarises many variables}
\usage{
lp_summarise(.data, ...)
}
\arguments{
\item{.data}{A tibble}

\item{...}{Select variables DPLYR style.
One or more unquoted expressions separated by commas. You can treat variable names like they are positions, so you can use expressions like x:y to select ranges of variables}
}
\value{
A tibble with summaries
}
\description{
This function cretes a summary table of multiple variables. It is unsing tidyevaluation
to gather the variables provided, and provide pre-defined summary statistics.
You can also supply grouped tibbles, to get summary statistics per group
}
\examples{
\dontrun{
mtcars \%>\%
group_by(vs) \%>\%
lp_summarise(mpg, cyl)
}
}
