% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_testRtest.R
\name{lp_testRtest}
\alias{lp_testRtest}
\title{Calculate the Test- Retest correlation among the variables}
\usage{
lp_testRtest(.test, .retest, .id = "id", .method = "pearson")
}
\arguments{
\item{.test}{The TEST dataframe}

\item{.retest}{The RE-TEST datafra,e}

\item{.id}{The id variable for the subjects (defaults to `id`)}

\item{.method}{The correlation method. "pearson","spearman", "kendall"? see \code{?corrr::correlate}}
}
\value{
A tibble of the correaltion coefficient per variable
}
\description{
This function takes two dataframes with the same col names and an id variable and calculates
the correlation within each column as a test-retest reliability index.
}
\details{
Checks whether the column names are the same, and if there is an `id` variable for the subjects
}
\examples{
test = data.frame(id = 1:10,
var_1 = rnorm(10),
var_2 = rnorm(10)
)

retest = data.frame(id = 1:10,
                var_1 = rnorm(10),
                var_2 = rnorm(10))


lp_testRtest(test, retest)
}
