% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_means.R
\name{table_means}
\alias{table_means}
\title{Craate a 2 by n, table of means and sd's}
\usage{
table_means(data, group, ...)
}
\arguments{
\item{data}{A dataframe, tibble}

\item{group}{The group variable. Must have 2 distince values}

\item{...}{The variables for the means to be calculated. DPlyr style}
}
\value{
A tibble
}
\description{
This function bla bla
}
\details{
Notes
}
\examples{
\dontrun{
mtcars \%>\% table_means(group = am, disp:wt)
}
}
