% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_t_test.R
\name{table_t_test}
\alias{table_t_test}
\title{A table of t-tests statistics across all the variables you input}
\usage{
table_t_test(data, group, ...)
}
\arguments{
\item{data}{A dataframe, tibble}

\item{group}{The grouping variable. Must have only two distinct values}

\item{...}{The variables to turn the tests across the group variable}
}
\value{
A tibble
}
\description{
A desctiption
}
\details{
Notes
}
\examples{
\dontrun{
mtcars \%>\% table_means(group = am, disp:wt)}
}
\seealso{
\code{\link{table_means}} for a t-test table
}
