% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_histogram.R
\name{facet_histogram}
\alias{facet_histogram}
\title{Faceted distribution of variables as histograms}
\usage{
facet_histogram(data, ..., statistic = c("mean", "median"))
}
\arguments{
\item{data}{The data}

\item{...}{The variables to plot. Unquoted, dplyr way}

\item{statistic}{One of "mean", "median". Defaults to "mean"}
}
\value{
A ggplot
}
\description{
Faceted distribution of variables as histograms
}
