% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_bar_chart.R
\name{lp_bar_chart}
\alias{lp_bar_chart}
\title{A bar chart with percentages}
\usage{
lp_bar_chart(data, var, .title = NULL)
}
\arguments{
\item{data}{The dataset}

\item{var}{The variable to plot}

\item{.title}{Optional. The title}
}
\value{
A ggplot object
}
\description{
A bar chart with percentages
}
