% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_bar_chart.R
\name{lp_bar_chart}
\alias{lp_bar_chart}
\title{A bar chart with percentages}
\usage{
lp_bar_chart(
  data,
  var,
  .pct_reorder = TRUE,
  .drop_levels = TRUE,
  .hjust = -0.2,
  .xlab = NULL,
  .ylab = NULL,
  .title = NULL,
  .bar_col = "#EE5C42"
)
}
\arguments{
\item{data}{The dataset}

\item{var}{The variable to plot}

\item{.pct_reorder}{Logical length 1. Reorder the bars based on value?}

\item{.drop_levels}{Logical length 1. Drop unused levels in the factor when counting ? i.e. do not show '0' count}

\item{.hjust}{Numeric length 1. The jhust for the value labels.}

\item{.xlab}{The x-axis label. Defaults to the variable name}

\item{.ylab}{The y-axis label. Defaults to the variable name}

\item{.title}{Optional title}

\item{.bar_col}{The colour of the bars}
}
\value{
A ggplot object
}
\description{
A bar chart with percentages
}
\examples{
lp_bar_chart(iris, Species)
}
