% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_by_target.R
\name{lp_scatter_by_target}
\alias{lp_scatter_by_target}
\title{Scatterplots of one variable to many others}
\usage{
lp_scatter_by_target(
  .data,
  target = NULL,
  ...,
  width = 25,
  .cor_method = "pearson",
  .smouth = "lm"
)
}
\arguments{
\item{.data}{The dataframe}

\item{target}{The target variable to plot scatter against all other vars}

\item{...}{The variables to use. DPLYR style. Supports \code{dplyr::everything()}}

\item{width}{The width of the labeller function in the strips. Defaults t0 25}

\item{.cor_method}{The correlation method. One of "pearson", "kendall", spearman". Defaults to "pearson"}

\item{.smouth}{The smooth method for the graphs. All option available in the ggplot2::geom_smouth()}
}
\value{
A ggplot object
}
\description{
Scatterplots of one variable to many others
}
\examples{
lp_scatter_by_target(mtcars, target = mpg, disp:wt)
}
