% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_by_target.R
\name{scatter_by_target}
\alias{scatter_by_target}
\title{Scatterplots of one variable to many others}
\usage{
scatter_by_target(.data, target = NULL, ..., width = 25)
}
\arguments{
\item{.data}{The dataframe}

\item{target}{The target variable to plot scatter against all other vars}

\item{...}{The variables to use. DPLYR style. Supports \code{dplyr::everything()}}

\item{width}{The width of the labeller function in the strips. Defaults t0 25}
}
\value{
A ggplot object
}
\description{
Scatterplots of one variable to many others
}
\examples{
scatter_by_target(mtcars, disp:wt, target = mpg)
}
