% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_means.R
\name{table_means}
\alias{table_means}
\title{Craate a table of means and sd's}
\usage{
table_means(.data, group, ..., add_totals = FALSE)
}
\arguments{
\item{.data}{A dataframe or tibble}

\item{group}{The group variable. Must be of type character or factor}

\item{...}{The variables for the means to be calculated. DPlyr style}

\item{add_totals}{Logical. Add the overall aggregations? Defaults to FALSE.}
}
\value{
A tibble
}
\description{
This function bla bla
}
\details{
Notes
}
\examples{
\dontrun{
mtcars \%>\% table_means(group = am, disp:wt)
}
}
